requirements.txt
Django==2.2.3
django-cors-headers==3.0.2
django-debug-toolbar==2.0
djangorestframework==3.9.4
pycodestyle==2.5.0
autopep8==1.4.4

